Proyecto/Servicios/Usuario/src/requirements.txt
Flask
mysql.connector-python